+++ backend/canisters/community/impl/src/model/groups_being_imported.rs
@@ -62,7 +62,7 @@ impl GroupsBeingImported {
             group.current_batch_started = None;
             group.error_message = None;
             group.bytes.extend_from_slice(bytes);
-            group.bytes.len() as u64 == group.total_bytes
+            group.is_complete()
         } else {
             false
         }
@@ -85,6 +85,10 @@ impl GroupsBeingImported {
 
     pub fn summaries(&self) -> Vec<GroupBeingImportedSummary> {
         self.groups.values().map(|g| g.into()).collect()
+    }
+
+    pub fn completed_imports(&self) -> Vec<ChatId> {
+        self.groups.iter().filter(|(_, g)| g.is_complete()).map(|(g, _)| *g).collect()
     }
 }
 
@@ -130,6 +134,10 @@ impl GroupBeingImported {
 
     pub fn is_default(&self) -> Timestamped<bool> {
         self.is_default.clone()
+    }
+
+    pub fn is_complete(&self) -> bool {
+        self.bytes.len() as u64 == self.total_bytes
     }
 }