+++ frontend/openchat-agent/src/services/community/candid/types.d.ts
@@ -362,7 +362,6 @@ export type CommunityPermissionRole = { 'Owners' : null } |
   { 'Members' : null };
 export interface CommunityPermissions {
   'create_public_channel' : CommunityPermissionRole,
-  'block_users' : CommunityPermissionRole,
   'update_details' : CommunityPermissionRole,
   'remove_members' : CommunityPermissionRole,
   'invite_users' : CommunityPermissionRole,
@@ -1121,7 +1120,6 @@ export interface NotificationEnvelope {
 }
 export interface OptionalCommunityPermissions {
   'create_public_channel' : [] | [CommunityPermissionRole],
-  'block_users' : [] | [CommunityPermissionRole],
   'update_details' : [] | [CommunityPermissionRole],
   'remove_members' : [] | [CommunityPermissionRole],
   'invite_users' : [] | [CommunityPermissionRole],