+++ frontend/openchat-client/src/utils/chat.ts
@@ -802,7 +802,7 @@ export function canChangePermissions(chat: ChatSummary): boolean {
         (chat.kind === "group_chat" || chat.kind === "channel") &&
         chat.membership !== undefined &&
         !chat.frozen &&
-        isPermitted(chat.membership.role, chat.permissions.changePermissions)
+        hasOwnerRights(chat.membership.role)
     );
 }
 
@@ -840,7 +840,7 @@ export function canBlockUsers(chat: ChatSummary): boolean {
         return (
             chat.public &&
             !chat.frozen &&
-            isPermitted(chat.membership.role, chat.permissions.blockUsers)
+            isPermitted(chat.membership.role, chat.permissions.removeMembers)
         );
     } else {
         return true;
@@ -852,7 +852,7 @@ export function canUnblockUsers(chat: ChatSummary): boolean {
         return (
             chat.public &&
             !chat.frozen &&
-            isPermitted(chat.membership.role, chat.permissions.blockUsers)
+            isPermitted(chat.membership.role, chat.permissions.removeMembers)
         );
     } else {
         return true;